[Setup] Move footer-related components to own folder
- for better organization/architecture
