+++ src/components/datagrid/_index.scss
@@ -2,7 +2,7 @@
 @import 'mixins';
 @import 'data_grid';
 @import 'body/header/data_grid_header_row';
-@import 'body/data_grid_footer_row';
+@import 'body/footer/data_grid_footer_row';
 @import 'body/header/data_grid_column_resizer';
 @import 'data_grid_data_row';
 @import 'controls/data_grid_toolbar';
