+++ src/components/datagrid/body/footer/data_grid_footer_row.tsx
@@ -8,9 +8,9 @@
 
 import classnames from 'classnames';
 import React, { forwardRef, memo, useContext } from 'react';
-import { EuiDataGridCell } from './data_grid_cell';
-import { DataGridCellPopoverContext } from './data_grid_cell_popover';
-import { EuiDataGridFooterRowProps } from '../data_grid_types';
+import { EuiDataGridCell } from '../data_grid_cell';
+import { DataGridCellPopoverContext } from '../data_grid_cell_popover';
+import { EuiDataGridFooterRowProps } from '../../data_grid_types';
 
 const EuiDataGridFooterRow = memo(
   forwardRef<HTMLDivElement, EuiDataGridFooterRowProps>(
